Add ability to set an object as public
No token is needed to get the object if it is public. Previously, we would
have to go into the AWS console and set an object as public if we wanted to
have an image on for example, the website. Now we can do it here with the click
of a button.