feat: Migrate from login to sso
Migrates the used login service from "login" to "sso". This required two
new backend dependencies:
- authlib
- pyjwt

This also required creating a new dev dockerfile which installs nginx,
so that nyckeln can be made available to the backend container.

Since sso can also return the hive keys of a user the dependency on Hive
was removed.