fix: Set accidentally removed cookies
So turns out that the frontend dependend on the "permissions" cookie
being set, which I foolishly removed in the last commit since they
seemed unused.

I also decided to keep the user cookie (I renamed it to "kth_id" to be
consistent), but I'm not sure where/if it's used.